refactor(credit): modularize retrieval_governance → retrieval/ package (Wave 1 PR #4)
Split 1326 LOC monolithic retrieval_governance.py into edgar-style package:
- models.py: exceptions, constants, thresholds, query maps (leaf)
- query_map.py: chapter-specialized query expansion
- evidence.py: gather_chapter_evidence + validate_provenance
- corpus.py: build_ic_corpus + ic_coverage_rerank
- saturation.py: enforce_evidence_saturation + build_retrieval_audit
- benchmarks.py: retrieve_market_benchmarks

Standard imports (no PEP 562 — stdlib only deps). structlog replaces
stdlib logging. Never-raises error contract documented. All callers
updated: deep_review.py, deep_review_corpus.py, memo_chapters.py.
