+++ backend/vertical_engines/credit/deep_review.py
@@ -357,12 +357,12 @@ def run_deal_deep_review_v4(
         load_cached_artifact_v4,
     )
     from ai_engine.portfolio.concentration_engine import compute_concentration
+    from vertical_engines.credit.critic import critique_intelligence
     from vertical_engines.credit.edgar import (
         build_edgar_multi_entity_context,
         extract_searchable_entities,
         fetch_edgar_multi_entity,
     )
-    from vertical_engines.credit.critic import critique_intelligence
     from vertical_engines.credit.ic_quant_engine import compute_quant_profile
     from vertical_engines.credit.memo_book_generator import generate_memo_book
     from vertical_engines.credit.memo_evidence_pack import (
@@ -470,7 +470,7 @@ def run_deal_deep_review_v4(
     from ai_engine.governance.policy_loader import (
         SEARCH_ENDPOINT as _SEARCH_ENDPOINT,
     )
-    from vertical_engines.credit.retrieval_governance import retrieve_market_benchmarks
+    from vertical_engines.credit.retrieval import retrieve_market_benchmarks
 
     _market_benchmarks: list[dict] = []
     _mb_chapters = ("ch03_exit", "ch08_returns", "ch09_downside", "ch12_peers")
@@ -1639,14 +1639,14 @@ async def async_run_deal_deep_review_v4(
         load_cached_artifact_v4,
     )
     from ai_engine.portfolio.concentration_engine import compute_concentration
+    from vertical_engines.credit.critic import (
+        classify_instrument_type,
+        critique_intelligence,
+    )
     from vertical_engines.credit.edgar import (
         build_edgar_multi_entity_context,
         extract_searchable_entities,
         fetch_edgar_multi_entity,
-    )
-    from vertical_engines.credit.critic import (
-        classify_instrument_type,
-        critique_intelligence,
     )
     from vertical_engines.credit.ic_quant_engine import compute_quant_profile
     from vertical_engines.credit.memo_book_generator import async_generate_memo_book
@@ -1738,7 +1738,7 @@ async def async_run_deal_deep_review_v4(
         from ai_engine.governance.policy_loader import (
             SEARCH_ENDPOINT as _SE,
         )
-        from vertical_engines.credit.retrieval_governance import retrieve_market_benchmarks
+        from vertical_engines.credit.retrieval import retrieve_market_benchmarks
         results: list[dict] = []
         chapters = ("ch03_exit", "ch08_returns", "ch09_downside", "ch12_peers")
