+++ docs/plans/2026-03-15-refactor-credit-vertical-modular-alignment-wave1-plan.md
@@ -204,7 +204,7 @@ kyc/
 
 ---
 
-### Phase 4: retrieval/ (PR #4)
+### Phase 4: retrieval/ (PR #4) ✅
 
 **Source:** `retrieval_governance.py` (1326 LOC)
